+++ src/embeddingUtil.ts
@@ -1,12 +1,16 @@
 import { getOpenaiApi } from './util';
 
-async function getEmbedding(text: string, engine="text-embedding-ada-002"): Promise<number[]> {
+async function getEmbedding(text: string, engine="text-embedding-ada-002"): Promise<number[] | undefined> {
   text = text.replace("\n", " ")
-  const embedding = await getOpenaiApi().createEmbedding({
-    model: engine,
-    input: text,
-  });
-  return embedding.data.data[0].embedding;
+  try {
+    const embedding = await getOpenaiApi().createEmbedding({
+      model: engine,
+      input: text,
+    });
+    return embedding.data.data[0].embedding;
+  } catch (error) {
+    console.log('createEmbeddingError', error);
+  }
 }
 
 export {