+++ src/searchServer/index.ts
@@ -13,14 +13,21 @@ type SearchParams = {
   lines?: number, // code lines
 }
 
+let prevTarget: string;
 app.get('/', async (req: Request, res) => {
-  const {keyword, model = 'text-embedding-ada-002', target = 'example', n = 3, lines} = req.query as unknown as SearchParams;
-  loadEnv(target);
+  const {keyword, model = 'text-embedding-ada-002', target = process.env.DEFAULT_TARGET, n = 3, lines} = req.query as unknown as SearchParams;
+  if (!target) {
+    return res.json({code: 401, result: null, err: new Error('target not specified'), message: 'target not specified'})
+  }
+  if (prevTarget !== target) {
+    loadEnv(target);
+    prevTarget = target;
+  }
   try {
     const searchResult = await searchFunction(keyword, n, lines);
     res.json({code: 200, result: searchResult, err: null});
   } catch (error) {
-    res.json({code: 401, result: null, err: error});
+    res.json({code: 401, result: null, err: error, message: error?.message});
   }
 })