+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 0.2.9
+### Features
+- search server add DEFAULT_TARGET, so you don't need pass target when you call search api
+- search cache
+
+### Fix
+- error catch process
+
 # 0.2.8
 
 ### Features